Add files via upload
Freely available version of CITCOM distributed as part of the CIDER 2006 meeting, from Peter van Keken.  This version is set up to compile on older Macintosh computers.  There is a known issue with this version of CITCOM related to the "malloc" commands -- the code sometimes refers to unallocated sectors of memory before they are allocated.  As a result, the code must be compiled with gcc3 and on a 32 bit computer.  One way around this is to modify and compile CITCOM on a virtual box running, for example, Ubuntu 12.01, and then run it natively on an Ubuntu machine running a modern version of Ubuntu (e.g., 16).  The executable will work even though you cannot re-compile.